Renaming splunk -> statistic.
