+++ dpppt-additionalinfo-backend/src/main/java/org/dpppt/additionalinfo/backend/ws/statistics/StatisticClient.java
@@ -1,7 +1,7 @@
-package org.dpppt.additionalinfo.backend.ws.splunk;
+package org.dpppt.additionalinfo.backend.ws.statistics;
 
 import org.dpppt.additionalinfo.backend.ws.model.statistics.Statistics;
 
-public interface SplunkClient {
+public interface StatisticClient {
     public Statistics getStatistics();
 }
